add homepage and create route for retro
